kmsgq-jenkins: use requested systems for SRU
Use a new s._request_systems variable to store the system explicitly listed in the SRU request.

The s.jd['system'] cannot be used as it will be updated when checking throught all the test cases.
The s._sut_pool cannot be used too as it has a more complex structure, not just a list.

This patch has been tested with:
  1. Submit a request without any systems specified.
  2. Submia a request with systems specified, including system that got blacklisted.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>