provisioning.py: remove lowlatency / kvm kernel installation in install_kernel_flavour()
Remove the installation part for lowlatency / kvm kernel in
install_kernel_flavour(), as the required package has already been addressed by
the pkg_name hack in kmsgq-jenkins.

Furthermore this simple linux-%s is not correct for kernels like HWE-edge, which
should be linux-lowlatency-hwe-16.04 instead, so it's safe to remove this.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>