lib/testsprops.py: pull ubuntu_ltp_syscalls out of sru-3 group
The ubuntu_ltp_syscalls test poses a risk to crash the system by causing
kernel oops / panic for an unfix issue.

Pull this test out for the sru-3 test group, and run it solely.

For those testing nodes that will run sru-3, they should all have an
ubuntu_ltp_syscalls test explictly listed.

Cloud testing is not affected as it's not creating a job group but creating
all the jobs individually.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>