Add dist-upgrade process for manual prov systems
The dist-upgrade process for manual provisioned systems
was removed in 857f1752.

Add it back along with an extra s.verify_target() check,
so that it is expected not to get executed on every jenkins
jobs for manual provisioned systems.

By doing this way, We don't need to manually upgrade the SUT
and restart those jobs failed with unmatched kernel version.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>