Improve the package string handling
The package passed from the test request is the source package name, not the
meta package name. Before we have this fixed in kteam-tools it is better to
compose the meta package name for all kernels (except linux-oem / linux-fips,
they are already the right package name to install).

Replace the 'pkg-name' to 'meta-pkg' to make it more straight forward and less
confusing with the 'package' passed from the test request.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>