lib/pdu.py: fetch the lab-systems-power.yaml locally
Not sure why we are hosting the lab-systems-power.yaml file on MaaS server,
it costs extra work to maintain a file out of tree there.

Move the lab-systems-power.yaml file to the lib directory, and change the pdu
script to make it read those information directly.

It will be easier to keep the yaml file up-to-date here.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>