lib/provisioning.py: retire the install_kernel_flavour function
This install_kernel_flavour() was designed for installing FIPS kernel, with
information provided in flavour-extras file.

With those information moved to ppa-extras and handled by enable_ppa() just
like how we deal with those ESM kernels, this function can be retired.

Packages for FIPS kernel will be installed by the install_kernel_meta_pkg(),
with meta package name linux-fips.

Therefore this change should not cause any trouble as long as we sent out
the test request in a proper format like what we did for ESM kernels.

Note: I didn't change this in lib/target.py, as it's for cloud and we're using
CKCT instead for cloud deployment now.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>